backend/requirement.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
cloudinary==1.36.0
python-dotenv==1.0.0
motor
pymongo